+++ routes/web.php
@@ -49,6 +49,7 @@ Route::middleware('throttle:booking-slots')->group(function () {
     Route::get('/book/{tenantSlug}/{locationSlug}', [PublicBookingController::class, 'show'])->name('booking.show');
     Route::get('/r/{tenantSlug}/{locationSlug}', [PublicBookingController::class, 'show']);
     Route::get('/book/{tenantSlug}/{locationSlug}/slots', [PublicBookingController::class, 'slots'])->name('booking.slots');
+    Route::get('/book/{tenantSlug}/{locationSlug}/floorplan', [PublicBookingController::class, 'floorplan'])->name('booking.floorplan');
     Route::get('/embed/{tenantSlug}/{locationSlug}.js', [PublicBookingController::class, 'embedScript'])->name('booking.embed');
     Route::get('/book/{tenantSlug}/{locationSlug}/events', [PublicEventController::class, 'index'])->name('events.index');
     Route::get('/book/{tenantSlug}/{locationSlug}/events/{eventSlug}', [PublicEventController::class, 'show'])->name('events.show');